fix sendMessage endpoint; improve public API security; improve telegram model; improve general runtime exception logging;

+++ src/main/kotlin/com/rocketraccoons/hyenas/cl4ptp/model/User.kt
@@ -6,6 +6,6 @@ import kotlinx.nosql.Id
 /**
  * Created by instu_000 on 3/4/2016.
  */
-data class User(val id: Int, val firstName: String, val lastName: String?, val username: String?, var isMe: Boolean) {
+data class User(val id: Long, val firstName: String, val lastName: String?, val username: String?, var isMe: Boolean) {
     val dbId: Id<String, Users>? = null
 }
